% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{air_passengers}
\alias{air_passengers}
\title{air_passengers data}
\format{
An object of class \code{data.frame} with 144 rows and 2 columns.
}
\source{
\code{datasets::AirPassengers}
}
\usage{
air_passengers
}
\description{
Data where the base \code{datasets::AirPassengers} monthly time
  series of international airline passenger numbers 1949-1960 was converted to
  a data frame using \code{tsbox::ts_df(datasets::AirPassengers)}.

\tabular{ll}{
  \strong{time} \tab Date with format 1949-01-01 \cr
  \strong{value} \tab numeric number of passengers in thousands
}
}
\keyword{datasets}
