% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{car_stats}
\alias{car_stats}
\title{car_stats data}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 3 rows and 6 columns.
}
\source{
\code{datasets::mtcars}
}
\usage{
car_stats
}
\description{
From \code{datasets::mtcars} produces statistics for N and the means of
"mpg", "hp", "wt", "disp" variables grouped by "cyl".

\tabular{ll}{
  \strong{N} \tab numeric of group's total observations \cr
  \strong{mean_mpg} \tab numeric of group's mean mpg \cr
  \strong{mean_hp} \tab numeric of group's mean horsepower \cr
  \strong{mean_wt} \tab numeric of group's mean weight \cr
  \strong{mean_disp} \tab numeric of group's mean displacment
}
}
\keyword{datasets}
