% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chick_weights}
\alias{chick_weights}
\title{chick_weights data}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 4 rows and 3 columns.
}
\source{
\code{datasets::ChickWeight}
}
\usage{
chick_weights
}
\description{
A dataframe with 4 observations of average chick weights.

\tabular{ll}{
 \strong{Diet} \tab numeric diet index \cr
 \strong{Weight} \tab numeric of average chick's weight \cr
 \strong{Label} \tab string label for the diet in percent of protein
}
}
\keyword{datasets}
