% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bar_plot.R
\name{create_bar_plot}
\alias{create_bar_plot}
\title{Function creates a bar plot.}
\usage{
create_bar_plot(
  df,
  aes_x,
  aes_y = NULL,
  aes_color = NULL,
  aes_fill = NULL,
  position = "stack",
  title = NULL,
  subtitle = NULL,
  center_titles = FALSE,
  x_title = aes_x,
  y_title = "Count",
  rot_x_tic_angle = 0,
  rot_y_tic_label = FALSE,
  bar_fill = NA,
  bar_color = "black",
  bar_alpha = 1,
  bar_size = 1,
  bar_width = NULL,
  palette_colors = NULL,
  x_limits = NULL,
  x_breaks = waiver(),
  x_labels = waiver(),
  x_date_breaks = waiver(),
  x_date_labels = waiver(),
  y_limits = NULL,
  y_major_breaks = waiver(),
  y_minor_breaks = waiver(),
  y_labels = waiver(),
  axis_text_size = 11,
  do_coord_flip = FALSE,
  order_bars = NULL,
  bar_labels = FALSE,
  bar_label_size = 4,
  bar_label_color = "black",
  show_major_grids = TRUE,
  show_minor_grids = TRUE,
  do_x_title = TRUE,
  do_y_title = TRUE,
  show_legend = TRUE,
  legend_pos = "top",
  silent_NA_warning = FALSE
)
}
\arguments{
\item{df}{The target data frame for the bar chart.}

\item{aes_x}{A string that sets the x axis discrete variable name from \code{df}. This variable must be a factor.}

\item{aes_y}{A string that sets the y axis numeric variable name from \code{df}.}

\item{aes_color}{A string that sets the variable name from \code{df} for the aesthetic mapping for color.}

\item{aes_fill}{A string that sets the variable name from \code{df} for the aesthetic mapping for fill.}

\item{position}{A string that sets the bar positions.  Acceptable values are \code{dodge}(side by side),
\code{identity}(overlap) or \code{stack}.}

\item{title}{A string that sets the overall title.}

\item{subtitle}{A string that sets the overall subtitle.}

\item{center_titles}{A logical which if \code{TRUE} centers both the \code{title} and \code{subtitle}.}

\item{x_title}{A string that sets the x axis title.}

\item{y_title}{A string that sets the y axis title.}

\item{rot_x_tic_angle}{A numeric that sets the angle of rotation for the x tic label. When x tic labels are long,
a value of 40 for this argument usually works well.}

\item{rot_y_tic_label}{A logical which if TRUE rotates the y tic labels 90 degrees for enhanced readability.}

\item{bar_fill}{A string that sets the fill color for the bars.}

\item{bar_color}{A string that sets the outline color for the bars.}

\item{bar_alpha}{A numeric that sets the alpha component to \code{bar_color}.}

\item{bar_size}{A numeric that sets the outline thickness of the bars.}

\item{bar_width}{A numeric that sets the width of the bars.}

\item{palette_colors}{A character vector to set the palette colors.}

\item{x_limits}{Depending on the class of \code{aes_x}, a numeric/Date/POSIXct 2 element vector that sets the minimum
and maximum for the x axis. Use NA to refer to the existing minimum and maximum..}

\item{x_breaks}{Depending on the class of \code{aes_x}, a numeric/Date/POSIXct vector or function that defines the
exact major tic locations along the x axis.}

\item{x_labels}{A character vector that defines the x axis tic labels. Vector must be the same length as
\code{x_breaks}.}

\item{x_date_breaks}{If the class of \code{aes_x} is Date/POSIXct, a string containing the number and date unit for
major breaks. \code{"1 years"}, \code{"4 sec"}, \code{"3 month"}, \code{"2 week"}.}

\item{x_date_labels}{If the class of \code{aes_x} is Date/POSIXct, a string containing the format codes, the
strftime format, for the date. Examples: \code{\%Y-\%m}, \code{\%Y/\%b/\%d}, \code{\%H-\%M-\%S}}

\item{y_limits}{A numeric 2 element vector that sets the minimum and maximum for the y axis.
Use NA to refer to the existing minimum and maximum.}

\item{y_major_breaks}{A numeric vector or function that defines the exact major tic locations along the y axis.}

\item{y_minor_breaks}{A numeric vector or function that defines the exact minor tic locations along the y axis.}

\item{y_labels}{A character vector with the same length as \code{y_major_breaks}, that labels the major tics.}

\item{axis_text_size}{A numeric that sets the font size along the axis'. Default is 11.}

\item{do_coord_flip}{A logical which if \code{TRUE} will flip the x and y axis'.}

\item{order_bars}{A string which will order the bars in a specific direction. Acceptable values are \dQuote{asc} or \dQuote{desc}}

\item{bar_labels}{A logical which if \code{TRUE} will label each bar with its value.}

\item{bar_label_size}{A numeric that sets the size of the bar labels}

\item{bar_label_color}{A string that sets the color of the bar labels}

\item{show_major_grids}{A logical that controls the appearance of major grids.}

\item{show_minor_grids}{A logical that controls the appearance of minor grids.}

\item{do_x_title}{A logical that controls the appearance of the x axis title.}

\item{do_y_title}{A logical that controls the appearance of the y axis title.}

\item{show_legend}{A logical that controls the appearance of the legend.}

\item{legend_pos}{A string that sets the legend position. Acceptable values are
\dQuote{top}, \dQuote{bottom}, \dQuote{left}, \dQuote{right}.}

\item{silent_NA_warning}{A logical that controls the appearance of a console warning when Na's
are removed.}
}
\value{
A plot object.
}
\description{
Function creates a ggplot2 based bar plot with optional scaling and ordering.
}
\author{
Rick Dean
}
