% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_box_plot.R
\name{create_box_plot}
\alias{create_box_plot}
\title{Function create a box plot.}
\usage{
create_box_plot(
  df,
  aes_x,
  aes_y,
  aes_color = NULL,
  aes_fill = NULL,
  title = NULL,
  subtitle = NULL,
  center_titles = FALSE,
  x_title = aes_x,
  y_title = aes_y,
  rot_x_tic_angle = 0,
  rot_y_tic_label = FALSE,
  box_fill = NA,
  box_color = "black",
  box_size = 1,
  box_alpha = 1,
  palette_colors = NULL,
  x_limits = NULL,
  x_breaks = waiver(),
  x_labels = waiver(),
  y_limits = NULL,
  y_major_breaks = waiver(),
  y_minor_breaks = waiver(),
  y_labels = waiver(),
  axis_text_size = 11,
  coord_flip = FALSE,
  do_x_title = TRUE,
  do_y_title = TRUE,
  show_legend = TRUE,
  legend_pos = "top",
  show_major_grids = TRUE,
  show_minor_grids = TRUE,
  show_outliers = FALSE,
  label_outlier_var = NULL,
  label_outlier_color = "black",
  ol_color = "black",
  ol_fill = "black",
  ol_size = 1.5,
  ol_shape = 19,
  ol_stroke = 0.5,
  ol_alpha = NULL,
  order_by_median = NULL,
  silent_NA_warning = FALSE
)
}
\arguments{
\item{df}{The target data frame for the density chart.}

\item{aes_x}{A string that sets the x axis factor variable name from \code{df}.}

\item{aes_y}{A string that sets the y axis variable name from \code{df}.}

\item{aes_color}{A string that sets the variable name from \code{df} for the aesthetic mapping for color.}

\item{aes_fill}{A string that sets the variable name from \code{df} for the aesthetic mapping for fill.}

\item{title}{A string that sets the overall title.}

\item{subtitle}{A string that sets the overall subtitle}

\item{center_titles}{A logical which if \code{TRUE} centers both the \code{title} and \code{subtitle}.}

\item{x_title}{A string that sets the x axis title.}

\item{y_title}{A string that sets the y axis title.}

\item{rot_x_tic_angle}{A numeric that sets the angle of rotation for the x tic labels. When x tic labels are long,
a value of 40 for this argument usually works well.}

\item{rot_y_tic_label}{A logical which if TRUE rotates the y tic labels 90 degrees for enhanced readability.}

\item{box_fill}{A string that sets the fill color for the box plot.}

\item{box_color}{A string that sets the color for the box plot.}

\item{box_size}{A numeric that sets the size of the box outline.}

\item{box_alpha}{A numeric that set the alpha component to \code{box_color}.}

\item{palette_colors}{A character vector to set the palette colors.}

\item{x_limits}{A character vector that defines possible values of the scale and their order..}

\item{x_breaks}{A character vector of x axis breaks or a function that takes the limits as input and returns breaks as output.}

\item{x_labels}{A character vector that defines the x axis tic labels. Vector must be the same length as \code{x_breaks}.}

\item{y_limits}{A numeric 2 element vector that sets the minimum and maximum for the y axis.
Use NA to refer to the existing minimum and maximum.}

\item{y_major_breaks}{A numeric vector or function that defines the exact major tic locations along the y axis.}

\item{y_minor_breaks}{A numeric vector or function that defines the exact minor tic locations along the y axis.}

\item{y_labels}{A character vector with the same length as \code{y_major_breaks}, that labels the major tics.}

\item{axis_text_size}{A numeric that sets the font size along the axis'. Default is 11.}

\item{coord_flip}{A logical which if TRUE will flip the x and y axis'.}

\item{do_x_title}{A logical that controls the appearance of the x axis title.}

\item{do_y_title}{A logical that controls the appearance of the y axis title.}

\item{show_legend}{A logical that controls the appearance of the legend.}

\item{legend_pos}{A string that sets the legend position. Acceptable values are
\dQuote{top}, \dQuote{bottom}, \dQuote{left}, \dQuote{right}.}

\item{show_major_grids}{A logical that controls the appearance of major grids.}

\item{show_minor_grids}{A logical that controls the appearance of minor grids.}

\item{show_outliers}{A logical which if TRUE will display the outliers.}

\item{label_outlier_var}{A string that sets the variable name from \code{df} whose value will to be displayed corresponding
to the \code{aes_y} outlier. If \code{show_outliers} is TRUE and this parameter is NULL, then the row number will be used as the label.}

\item{label_outlier_color}{A string that sets the color of outlier labels.}

\item{ol_color}{A string that sets the outlier color.}

\item{ol_fill}{A string that sets the outlier fill.}

\item{ol_size}{A numeric that set the outlier size.}

\item{ol_shape}{A string that set the outlier shape.}

\item{ol_stroke}{A numeric that sets the outlier shape line width.}

\item{ol_alpha}{A numeric that sets the outlier alpha for color.}

\item{order_by_median}{A string which will order the plot of \code{aes_x} factor levels/categories by the \code{aes_y} medians.
Acceptable values are \dQuote{asc} or \dQuote{desc}}

\item{silent_NA_warning}{A logical that controls the appearance of a console warning when Na's
are removed.}
}
\value{
A plot object.
}
\description{
Function creates a ggplot2 based box plot with optional scaling, ordering and outlier viewing.
}
\author{
Rick Dean
}
