% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_density_plot.R
\name{create_density_plot}
\alias{create_density_plot}
\title{Function creates a density plot.}
\usage{
create_density_plot(
  df,
  aes_x,
  bw = "nrd0",
  adjust = 1,
  kernel = "gaussian",
  n = 512,
  na.rm = TRUE,
  aes_color = NULL,
  aes_fill = NULL,
  position = "identity",
  title = NULL,
  subtitle = NULL,
  center_titles = FALSE,
  x_title = aes_x,
  y_title = "Density",
  rot_x_tic_angle = 0,
  rot_y_tic_label = FALSE,
  density_size = 1,
  density_color = "black",
  density_fill = "gray",
  density_alpha = 0.4,
  palette_colors = NULL,
  x_limits = NULL,
  x_major_breaks = waiver(),
  x_minor_breaks = waiver(),
  x_labels = waiver(),
  y_limits = NULL,
  y_major_breaks = waiver(),
  y_minor_breaks = waiver(),
  y_labels = waiver(),
  axis_text_size = 11,
  do_x_title = TRUE,
  do_y_title = TRUE,
  show_major_grids = TRUE,
  show_minor_grids = TRUE,
  plot_obs = FALSE,
  plot_obs_len = 0.02,
  plot_obs_color = "black",
  plot_obs_jitter = FALSE,
  show_legend = TRUE,
  legend_pos = "top",
  cum_prob = NULL,
  area_colors = c("gray", "green", "gray"),
  area_quantile_line_color = "red"
)
}
\arguments{
\item{df}{A data frame that contains a numeric vector from which to estimate the KDE along with possible variables
for setting \code{aes_color} and \code{aes_fill} (see below).}

\item{aes_x}{The variable from \code{df} from which to estimate the KDE.}

\item{bw}{A string or numeric that sets the smoothing bandwidth to be used with the KDE function.}

\item{adjust}{A numeric that adjusts \code{bw} since the actual bandwidth is computed as \code{adjust*bw}.}

\item{kernel}{A string that set the type of Kernel Density Estimation (KDE). Acceptable values are "gaussian",
"rectangular", "triangular", "epanechnikov", "biweight", "cosine" or "optcosine".}

\item{n}{The number of equally spaced points at which the density is to be estimated. This should be a power of two.}

\item{na.rm}{A logical which if \code{TRUE}, missing values are removed from \code{df}. If \code{FALSE} any missing
values cause an error.}

\item{aes_color}{A string that sets the variable name from \code{df} for the aesthetic mapping for color.}

\item{aes_fill}{A string that sets the variable name from \code{df} for the aesthetic mapping for fill.}

\item{position}{A string that sets the position.  Acceptable values are \code{identity} which overlays or
\code{stack} which stacks.}

\item{title}{A string that sets the overall title.}

\item{subtitle}{A string that sets the overall subtitle.}

\item{center_titles}{A logical which if \code{TRUE} centers both the \code{title} and \code{subtitle}.}

\item{x_title}{A string that sets the x axis title.}

\item{y_title}{A string that sets the y axis title.}

\item{rot_x_tic_angle}{A numeric that sets the angle of rotation for the x tic labels. When x tic labels are long,
a value of 40 for this argument usually works well.}

\item{rot_y_tic_label}{A logical which if TRUE rotates the y tic labels 90 degrees for enhanced readability.}

\item{density_size}{A numeric that sets the density line width.}

\item{density_color}{A string that sets the color for the density line.}

\item{density_fill}{A string that sets the color for the density fill.}

\item{density_alpha}{A numeric that sets the alpha value for \code{density_fill}.}

\item{palette_colors}{A character vector to set the palette colors.}

\item{x_limits}{A numeric 2 element vector or function that sets the minimum and maximum for the x axis. Use NA to
refer to the existing minimum and maximum.}

\item{x_major_breaks}{A numeric vector or function that sets the major tic locations along the x axis.}

\item{x_minor_breaks}{A numeric vector or function that sets the minor tic locations along the x axis.}

\item{x_labels}{A character vector or function giving x axis tic labels.  Must be the same length as \code{x_breaks}.}

\item{y_limits}{A numeric 2 element vector or function that sets the minimum and maximum for the y axis.  Use NA to
refer to the existing minimum and maximum.}

\item{y_major_breaks}{A numeric vector or function that sets the major tic locations along the y axis.}

\item{y_minor_breaks}{A numeric vector or function that sets the minor tic locations along the y axis.}

\item{y_labels}{A character vector or function giving y axis tic labels.  Must be the same length as \code{y_breaks}.}

\item{axis_text_size}{A numeric that sets the font size along the axis'. Default is 11.}

\item{do_x_title}{A logical that controls the appearance of the x axis title.}

\item{do_y_title}{A logical that controls the appearance of the y axis title.}

\item{show_major_grids}{A logical that controls the appearance of major grids.}

\item{show_minor_grids}{A logical that controls the appearance of minor grids.}

\item{plot_obs}{A logical which if \code{TRUE} plots a line for each observation along the axis margin.}

\item{plot_obs_len}{A numeric that sets the length of the \code{plot_obs} lines.}

\item{plot_obs_color}{A string that sets the color of the \code{plot_obs} lines.}

\item{plot_obs_jitter}{A logical which if \code{TRUE} will add a slight horizontal adjustment to overlapping observations.}

\item{show_legend}{A logical that controls the appearance of the legend.}

\item{legend_pos}{A string that sets the legend position. Acceptable values are
\dQuote{top}, \dQuote{bottom}, \dQuote{left}, \dQuote{right}.}

\item{cum_prob}{A one or two element numeric vector for defining cumulative probabilities which divide the density plot area. The
values are probabilities with values within \link{0.0, 1.0}.  If for example with a cumulative probability \code{cum_prob} of .95, then the density is divided
into two probability areas of 95\% and 5\% with the dividing quantile value labeled. Similarly if two cumulative probabilities are submitted,
then three areas are defined.}

\item{area_colors}{A string vector that sets the color of each area defined by \code{cum_prob}.}

\item{area_quantile_line_color}{A string that sets the vertical line color at quantile locations that divide the areas defined
by \code{cum_prob}.}
}
\value{
A ggplot2 plot object.
}
\description{
Function creates a ggplot2 based density plot with options for scaling, shading probability areas,
and plotting observation locations. The function's density arguments mirror most of the arguments available
from \code{stats::density()} for the Kernel Density Estimation (KDE). See the \code{stats::density()}'s help page
for more information.
}
\author{
Rick Dean
}
