% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_density_ridge_plot.R
\name{create_density_ridge_plot}
\alias{create_density_ridge_plot}
\title{Function creates a vertical stack of density plots.}
\usage{
create_density_ridge_plot(
  df,
  variables,
  plot_heights = NULL,
  bw = "nrd0",
  adjust = 1,
  kernel = "gaussian",
  n = 512,
  na.rm = TRUE,
  title = NULL,
  subtitle = NULL,
  x_title = NULL,
  rot_x_tic_angle = 0,
  density_size = 1,
  density_color = "black",
  density_fill = "gray",
  density_alpha = 0.4,
  x_limits = NULL,
  x_major_breaks = waiver(),
  x_minor_breaks = waiver(),
  x_labels = waiver(),
  y_limits = NULL,
  y_major_breaks = waiver(),
  y_minor_breaks = waiver(),
  y_show_axis = FALSE,
  axis_text_size = 11,
  display_plot = TRUE
)
}
\arguments{
\item{df}{The source data frame from which the densities are plotted.}

\item{variables}{A vector that names the x axis variables from \code{df} for plotting their densities.}

\item{plot_heights}{A vector that sets the plot height in inches for each variable in \code{variables}.
If this argument is NULL, then each plot will be 1.2 inches in height.}

\item{bw}{A string or numeric that sets the smoothing bandwidth to be used with the KDE function.}

\item{adjust}{A numeric that adjusts \code{bw} since the actual bandwidth is computed as \code{adjust*bw}.}

\item{kernel}{A string that set the type of Kernel Density Estimation (KDE). Acceptable values are "gaussian",
"rectangular", "triangular", "epanechnikov", "biweight", "cosine" or "optcosine".}

\item{n}{The number of equally spaced points at which the density is to be estimated. This should be a power of two.}

\item{na.rm}{A logical which if \code{TRUE}, missing values are removed from \code{df}. If \code{FALSE} any missing
values cause an error.}

\item{title}{A string that sets the overall title.}

\item{subtitle}{A string that sets the overall subtitle.}

\item{x_title}{A string that sets the x axis title. If NULL (the default)  then the x axis title does not appear.}

\item{rot_x_tic_angle}{A numeric that sets the angle of rotation for the x tic labels. When x tic labels are long,
a value of 40 for this argument usually works well.}

\item{density_size}{A numeric that sets the density line width.}

\item{density_color}{A string that sets the color for the density line.}

\item{density_fill}{A string that sets the color for the density fill.}

\item{density_alpha}{A numeric that sets the alpha value for \code{density_fill}.}

\item{x_limits}{A numeric 2 element vector or function that sets the minimum and maximum for the x axis. Use NA to
refer to the existing minimum and maximum.}

\item{x_major_breaks}{A numeric vector or function that sets the major tic locations along the x axis.}

\item{x_minor_breaks}{A numeric vector or function that sets the minor tic locations along the x axis.}

\item{x_labels}{A character vector or function giving x axis tic labels.  Must be the same length as \code{x_breaks}.}

\item{y_limits}{A numeric 2 element vector or function that sets the minimum and maximum for the y axis.  Use NA to
refer to the existing minimum and maximum.}

\item{y_major_breaks}{A numeric vector or function that sets the major tic locations along the y axis.}

\item{y_minor_breaks}{A numeric vector or function that sets the minor tic locations along the y axis.}

\item{y_show_axis}{A logical which if TRUE will display the y-axis density for each variable.}

\item{axis_text_size}{A numeric that sets the font size along the axis'. Default is 11.}

\item{display_plot}{A logical that if TRUE will display the plot.}
}
\value{
A TableGrob object if \code{display_plot} is FALSE.
}
\description{
Function creates ggplot2 based density plots stacked vertically (also known as ridge plots).
The function's density arguments mirror most of the arguments available
from \code{\link{density}} for the Kernel Density Estimation (KDE). See the \code{\link{density}} help page
for more information.
}
\author{
Rick Dean
}
