% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_heatmap.R
\name{create_heatmap}
\alias{create_heatmap}
\title{Function produces a heatmap that shows magnitude as an array of cells in two dimensions.}
\usage{
create_heatmap(
  df,
  aes_x,
  aes_y,
  aes_label,
  aes_fill = NULL,
  aes_size = NULL,
  aes_width = NULL,
  aes_height = NULL,
  tile_sz = 1.1,
  tile_color = "white",
  tile_fill = "white",
  tile_square = FALSE,
  label_sz = 11,
  label_color = "white",
  label_fontface = "plain",
  label_alpha = 1,
  title = NULL,
  subtitle = NULL,
  center_titles = FALSE,
  x_title = aes_x,
  y_title = aes_y,
  rot_x_tic_angle = 0,
  rot_y_tic_label = FALSE,
  x_limits = NULL,
  x_major_breaks = waiver(),
  x_labels = waiver(),
  x_major_date_breaks = waiver(),
  x_date_labels = waiver(),
  y_limits = NULL,
  y_major_breaks = waiver(),
  y_labels = waiver(),
  axis_text_size = 11,
  show_legend = TRUE,
  legend_pos = "top",
  silent_NA_warning = FALSE
)
}
\arguments{
\item{df}{The target data frame from which the heatmap is plotted.}

\item{aes_x}{A string that sets the x axis variable name from \code{df}. Is a required discrete numeric,
Date/POSIXct variable.}

\item{aes_y}{A string that sets the y axis variable name from \code{df}. Is a required discrete numeric.}

\item{aes_label}{A string that sets the variable name from \code{df} as the required source for the tile's text.}

\item{aes_fill}{A string that sets the discrete/continuous variable name from \code{df} for the aesthetic mapping for fill.}

\item{aes_size}{A string that sets the variable name from \code{df} for the aesthetic mapping for size.}

\item{aes_width}{A string that sets the variable name from \code{df} for the aesthetic mapping for width.}

\item{aes_height}{A string that sets the variable name from \code{df} for the aesthetic mapping for height.}

\item{tile_sz}{A numeric that sets the tile's size width in millimeters.}

\item{tile_color}{A numeric that sets the tile's border color.}

\item{tile_fill}{A string that sets the fill color for the tile.}

\item{tile_square}{A logical which if TRUE makes the tiles square instead of rectangular.}

\item{label_sz}{A numeric that sets the size of the label.}

\item{label_color}{A string that sets the label's color.}

\item{label_fontface}{A string that sets the label's font face. Acceptable values are \dQuote{plain}, \dQuote{bold},
\dQuote{italic}, \dQuote{bold.italic}.}

\item{label_alpha}{A numeric that sets the label's alpha value.}

\item{title}{A string that sets the plot title.}

\item{subtitle}{A string that sets the plot subtitle.}

\item{center_titles}{A logical which if \code{TRUE} centers both the \code{title} and \code{subtitle}.}

\item{x_title}{A string that sets the x axis title.}

\item{y_title}{A string that sets the y axis title.}

\item{rot_x_tic_angle}{A numeric that sets the angle of rotation for the x tic labels. When x tic labels are long,
a value of 40 for this argument usually works well.}

\item{rot_y_tic_label}{A logical which if TRUE rotates the y tic labels 90 degrees for enhanced readability.}

\item{x_limits}{Depending on the class of \code{aes_x}, a numeric/Date/POSIXct 2 element vector that sets the minimum
and maximum for the x axis. Use NA to refer to the existing minimum and maximum.}

\item{x_major_breaks}{Depending on the class of \code{aes_x}, a numeric/Date/POSIXct vector or function that defines
the exact major tic locations along the x axis.}

\item{x_labels}{A character vector with the same length as \code{x_major_breaks}, that labels the major tics.}

\item{x_major_date_breaks}{If the class of \code{aes_x} is Date/POSIXct, a string containing the number and date unit
for major breaks. \code{"1 year"}, \code{"4 sec"}, \code{"3 month"}, \code{"2 week"}.}

\item{x_date_labels}{If the class of \code{aes_x} is Date/POSIXct, a string containing the format codes, the strftime
format, for the date.
Examples: \code{\%Y-\%m}, \code{\%Y/\%b/\%d}, \code{\%H-\%M-\%S}}

\item{y_limits}{A numeric 2 element vector that sets the minimum and  maximum for the y axis.
Use \code{NA} to refer to the existing minimum and maximum.}

\item{y_major_breaks}{A numeric vector or function that defines the exact major tic locations along the y axis.}

\item{y_labels}{A character vector with the same length as \code{y_major_breaks}, that labels the major tics.}

\item{axis_text_size}{A numeric that sets the font size along the axis'. Default is 11.}

\item{show_legend}{A logical that controls the appearance of the legend.}

\item{legend_pos}{A string that sets the legend position. Acceptable values are
\dQuote{top}, \dQuote{bottom}, \dQuote{left}, \dQuote{right}.}

\item{silent_NA_warning}{A logical that controls the appearance of a console warning when Na's
are removed.}
}
\value{
A plot object
}
\description{
Function returns a ggplot2 x/y plot of a pair of discrete variables. Instead of a point, a
rectangle/cell/tile is located for each x/y value with the cell itself being colored or sized depending on the value
of a third associated discrete or continuous variable.
}
\author{
Rick Dean
}
