% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_histogram_plot.R
\name{create_histogram_plot}
\alias{create_histogram_plot}
\title{Function creates a histogram.}
\usage{
create_histogram_plot(
  df,
  aes_x,
  aes_color = NULL,
  aes_fill = NULL,
  position = "stack",
  title = NULL,
  subtitle = NULL,
  center_titles = FALSE,
  x_title = aes_x,
  y_title = "Count",
  rot_x_tic_angle = 0,
  rot_y_tic_label = FALSE,
  bins = 20,
  binwidth = NULL,
  bin_breaks = NULL,
  bar_fill = NA,
  bar_color = "black",
  bar_alpha = 0.4,
  bar_size = 1,
  palette_colors = NULL,
  x_limits = NULL,
  x_major_breaks = waiver(),
  x_minor_breaks = waiver(),
  x_labels = waiver(),
  y_limits = NULL,
  y_major_breaks = waiver(),
  y_minor_breaks = waiver(),
  y_labels = waiver(),
  axis_text_size = 11,
  do_coord_flip = FALSE,
  bar_labels = FALSE,
  bar_label_size = 6,
  bar_label_color = "black",
  plot_obs = FALSE,
  plot_obs_len = 0.02,
  plot_obs_color = "black",
  plot_obs_jitter = FALSE,
  do_x_title = TRUE,
  do_y_title = TRUE,
  show_major_grids = TRUE,
  show_minor_grids = TRUE,
  show_legend = TRUE,
  legend_pos = "top",
  silent_NA_warning = FALSE
)
}
\arguments{
\item{df}{The target data frame for the bar chart.}

\item{aes_x}{A string that sets the x axis continuous variable name from \code{df}.}

\item{aes_color}{A string that sets the variable name from \code{df} for the aesthetic mapping for color.}

\item{aes_fill}{A string that sets the variable name from \code{df} for the aesthetic mapping for fill.}

\item{position}{A string that sets the bar positions.  Acceptable values are \code{dodge, dodge2}(side by side), \code{identity}(overlap) or \code{stack}.}

\item{title}{A string that sets the overall title.}

\item{subtitle}{A string that sets the overall subtitle.}

\item{center_titles}{A logical which if \code{TRUE} centers both the \code{title} and \code{subtitle}.}

\item{x_title}{A string that sets the x axis title.}

\item{y_title}{A string that sets the y axis title.}

\item{rot_x_tic_angle}{A numeric that sets the angle of rotation for the x tic labels. When x tic labels are long,
a value of 40 for this argument usually works well.}

\item{rot_y_tic_label}{A logical which if TRUE rotates the y tic labels 90 degrees for enhanced readability.}

\item{bins}{An integer that sets the number of bins for the histogram. Is overridden by \code{binwidth}. Default is 20.}

\item{binwidth}{A numeric that sets the bin width of the histogram in units of \code{aes_x}.  If \code{aes_x} is a
date variable then \code{binwidth} is the number of days and if a time variable then \code{binwidth} is
the number of seconds.}

\item{bin_breaks}{A numeric vector giving the bin boundaries. Overrides \code{binwidth} and \code{bins}}

\item{bar_fill}{A string that sets the fill color for the bars.}

\item{bar_color}{A string that sets the outline color for the bars.}

\item{bar_alpha}{A numeric that set the alpha component to \code{bar_color}.}

\item{bar_size}{A numeric that sets the outline thickness of the bars.}

\item{palette_colors}{A character vector to set the palette colors.}

\item{x_limits}{A numeric 2 element vector or function that sets the minimum and maximum for the x axis.  Use NA to
refer to the existing minimum and maximum.}

\item{x_major_breaks}{A numeric vector or function that sets the major tic locations along the x axis.}

\item{x_minor_breaks}{A numeric vector or function that sets the minor tic locations along the x axis.}

\item{x_labels}{A character vector or function giving x axis tic labels.  Must be the same length as \code{x_breaks}.}

\item{y_limits}{A numeric 2 element vector or function that sets the minimum and maximum for the y axis.  Use NA to
refer to the existing minimum and maximum.}

\item{y_major_breaks}{A numeric vector or function that sets the major tic locations along the y axis.}

\item{y_minor_breaks}{A numeric vector or function that sets the minor tic locations along the y axis.}

\item{y_labels}{A character vector or function giving y axis tic labels.  Must be the same length as \code{y_breaks}.}

\item{axis_text_size}{A numeric that sets the font size along the axis'. Default is 11.}

\item{do_coord_flip}{A logical which if TRUE will flip the x and y axis'.}

\item{bar_labels}{A logical which if \code{TRUE} will label each bar with its value.}

\item{bar_label_size}{A numeric that sets the size of the bar labels}

\item{bar_label_color}{A string that sets the color of the bar labels}

\item{plot_obs}{A logical which if \code{TRUE} plots a line for each observation along the axis margin.}

\item{plot_obs_len}{A numeric that sets the length of the \code{plot_obs} lines.}

\item{plot_obs_color}{A string that sets the color of the \code{plot_obs} lines.}

\item{plot_obs_jitter}{A logical which if \code{TRUE} will add a slight horizontal adjustment to overlapping observations.}

\item{do_x_title}{A logical that controls the appearance of the x axis title.}

\item{do_y_title}{A logical that controls the appearance of the y axis title.}

\item{show_major_grids}{A logical that controls the appearance of major grids.}

\item{show_minor_grids}{A logical that controls the appearance of minor grids.}

\item{show_legend}{A logical that controls the appearance of the legend.}

\item{legend_pos}{A string that sets the legend position. Acceptable values are
\dQuote{top}, \dQuote{bottom}, \dQuote{left}, \dQuote{right}.}

\item{silent_NA_warning}{A logical that controls the appearance of a console warning when Na's
are removed.}
}
\value{
A plot object.
}
\description{
Function plots a ggplot2 based histogram with options for scaling and viewing observation locations.
}
\author{
Rick Dean
}
