% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_scatter_plot.R
\name{create_scatter_plot}
\alias{create_scatter_plot}
\title{Function produces a ggplot2 based scatter plot.}
\usage{
create_scatter_plot(
  df,
  aes_x,
  aes_y,
  aes_color = NULL,
  aes_fill = NULL,
  aes_size = NULL,
  aes_linetype = NULL,
  aes_label = NULL,
  position = position_jitter(width = 0, height = 0),
  title = NULL,
  subtitle = NULL,
  center_titles = FALSE,
  x_title = aes_x,
  y_title = aes_y,
  rot_x_tic_angle = 0,
  rot_y_tic_label = FALSE,
  x_limits = NULL,
  x_major_breaks = waiver(),
  x_minor_breaks = waiver(),
  x_labels = waiver(),
  x_major_date_breaks = waiver(),
  x_minor_date_breaks = waiver(),
  x_date_labels = waiver(),
  x_log10 = FALSE,
  y_limits = NULL,
  y_major_breaks = waiver(),
  y_minor_breaks = waiver(),
  y_labels = waiver(),
  y_log10 = FALSE,
  axis_text_size = 11,
  show_pts = TRUE,
  pts_fill = "white",
  pts_shape = 21,
  pts_stroke = 1,
  pts_color = "black",
  pts_line_alpha = 1,
  pts_size = 1,
  palette_colors = NULL,
  palette_linetypes = NULL,
  connect = FALSE,
  line_size = 1,
  line_color = "black",
  connect_linetype = "solid",
  do_x_title = TRUE,
  do_y_title = TRUE,
  show_major_grids = TRUE,
  show_minor_grids = TRUE,
  show_legend = TRUE,
  legend_pos = "top",
  bold_y = NULL,
  silent_NA_warning = FALSE
)
}
\arguments{
\item{df}{The target data frame from which the scatter points are plotted.}

\item{aes_x}{A string that sets the x axis variable name from \code{df}. Is a required discrete/continuous numeric,
Date/POSIXct variable.}

\item{aes_y}{A string that sets the y axis variable name from \code{df}. Is a required discrete/continuous numeric.}

\item{aes_color}{A string that sets the variable name from \code{df} for the aesthetic mapping for color.}

\item{aes_fill}{A string that sets the variable name from \code{df} for the aesthetic mapping for fill.}

\item{aes_size}{A string that sets the variable name from \code{df} for the aesthetic mapping for size.}

\item{aes_linetype}{A string that sets the variable name from \code{df} for the aesthetic mapping for linetype.}

\item{aes_label}{A string that sets the variable name from \code{df} for the aesthetic mapping for labeling. If
labeling points then the package \code{ggrepel} is required.}

\item{position}{A string or function that does a slight adjustment to overlapping points.  Typical values are
\code{jitter} or \code{position_jitter(width = 0.1, height = 0.1)}.}

\item{title}{A string that sets the plot title.}

\item{subtitle}{A string that sets the plot subtitle.}

\item{center_titles}{A logical which if \code{TRUE} centers both the \code{title} and \code{subtitle}.}

\item{x_title}{A string that sets the x axis title.}

\item{y_title}{A string that sets the y axis title.}

\item{rot_x_tic_angle}{A numeric that sets the angle of rotation for the x tic labels. When x tic labels are long,
a value of 40 for this argument usually works well.}

\item{rot_y_tic_label}{A logical which if TRUE rotates the y tic labels 90 degrees for enhanced readability.}

\item{x_limits}{Depending on the class of \code{aes_x}, a numeric/Date/POSIXct 2 element vector that sets the minimum
and maximum for the x axis. Use NA to refer to the existing minimum and maximum.}

\item{x_major_breaks}{Depending on the class of \code{aes_x}, a numeric/Date/POSIXct vector or function that defines
the exact major tic locations along the x axis.}

\item{x_minor_breaks}{Depending on the class of \code{aes_x}, a numeric/Date/POSIXct vector or function that defines
the exact minor tic locations along the x axis.}

\item{x_labels}{A character vector with the same length as \code{x_major_breaks}, that labels the major tics.}

\item{x_major_date_breaks}{If the class of \code{aes_x} is Date/POSIXct, a string containing the number and date unit
for major breaks. \code{"1 year"}, \code{"4 sec"}, \code{"3 month"}, \code{"2 week"}.}

\item{x_minor_date_breaks}{If the class of \code{aes_x} is Date/POSIXct, a string containing the number and date unit
for minor breaks.}

\item{x_date_labels}{If the class of \code{aes_x} is Date/POSIXct, a string containing the format codes, the strftime
format, for the date.
Examples: \code{\%Y-\%m}, \code{\%Y/\%b/\%d}, \code{\%H-\%M-\%S}}

\item{x_log10}{A logical which if \code{TRUE} will use a log10 scale for the x axis.}

\item{y_limits}{A numeric 2 element vector that sets the minimum and  maximum for the y axis.
Use \code{NA} to refer to the existing minimum and maximum.}

\item{y_major_breaks}{A numeric vector or function that defines the exact major tic locations along the y axis.}

\item{y_minor_breaks}{A numeric vector or function that defines the exact minor tic locations along the y axis.}

\item{y_labels}{A character vector with the same length as \code{y_major_breaks}, that labels the major tics.}

\item{y_log10}{A logical which if \code{TRUE} will use a log10 scale for the y axis.}

\item{axis_text_size}{A numeric that sets the font size along the axis'. Default is 11.}

\item{show_pts}{A logical which if FALSE will plot only the lines if \code{connect} is TRUE.}

\item{pts_fill}{A string that sets the fill color of the points.}

\item{pts_shape}{A numeric integer that sets the shape of the points. Typical values are 21 \dQuote{circle},
22 \dQuote{square}, 23 \dQuote{diamond}, 24 \dQuote{up triangle}, 25 \dQuote{down triangle}.}

\item{pts_stroke}{A numeric that sets the drawing width for a point shape.}

\item{pts_color}{A string that sets the color of the points.}

\item{pts_line_alpha}{A numeric value that sets the alpha level of points and connected line.}

\item{pts_size}{A numeric value that sets the size of the points.}

\item{palette_colors}{A string vector to set the aes mapping of colors which changes their appearance in the
order given.}

\item{palette_linetypes}{A string vector to set the aes mapping of linetypes which changes their appearance in
the order given.}

\item{connect}{A logical which if \code{TRUE} then points will be connected with a line.}

\item{line_size}{A numeric value that sets the thickness of lines if \code{connect} is TRUE.}

\item{line_color}{A string that sets the color of the lines if \code{connect} is TRUE.}

\item{connect_linetype}{A string that sets line type \code{twodash, solid, longdash, dotted, dotdash,
dashed, blank} if \code{connect} is TRUE.}

\item{do_x_title}{A logical that controls the appearance of the x axis title.}

\item{do_y_title}{A logical that controls the appearance of the y axis title.}

\item{show_major_grids}{A logical that controls the appearance of major grids.}

\item{show_minor_grids}{A logical that controls the appearance of minor grids.}

\item{show_legend}{A logical that controls the appearance of the legend.}

\item{legend_pos}{A string that sets the legend position. Acceptable values are
\dQuote{top}, \dQuote{bottom}, \dQuote{left}, \dQuote{right}.}

\item{bold_y}{A numeric that plots a bold horizontal line at this y value intercept.}

\item{silent_NA_warning}{A logical that controls the appearance of a console warning when Na's
are removed.}
}
\value{
A plot object
}
\description{
Function returns a ggplot2 plot object of x/y scatter points/lines. Options are provided for scaling and
line connection.
}
\author{
Rick Dean
}
