% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_stick_plot.R
\name{create_stick_plot}
\alias{create_stick_plot}
\title{Function produces a ggplot2 based "stick" plot.}
\usage{
create_stick_plot(
  df,
  base_val = 0,
  aes_x = NULL,
  aes_y = NULL,
  aes_color = NULL,
  title = NULL,
  subtitle = NULL,
  center_titles = FALSE,
  x_title = aes_x,
  y_title = aes_y,
  rot_x_tic_angle = 0,
  rot_y_tic_label = FALSE,
  x_limits = NULL,
  x_major_breaks = waiver(),
  x_labels = waiver(),
  x_major_date_breaks = waiver(),
  x_date_labels = waiver(),
  y_limits = NULL,
  y_major_breaks = waiver(),
  y_minor_breaks = waiver(),
  y_labels = waiver(),
  axis_text_size = 11,
  line_color = "black",
  line_size = 0.8,
  line_alpha = 1,
  palette_colors = NULL,
  do_x_title = TRUE,
  do_y_title = TRUE,
  show_major_grids = TRUE,
  show_minor_grids = TRUE,
  show_legend = TRUE,
  legend_pos = "top",
  bold_y = NULL,
  silent_NA_warning = FALSE
)
}
\arguments{
\item{df}{The target data frame from which the \dQuote{stick} lines are drawn.}

\item{base_val}{A numeric that sets the base value from which the \dQuote{stick} originates.
The default value is 0.}

\item{aes_x}{A string that sets the x axis variable name from \code{df}.  Can be a numeric/Date/POSIXct variable.}

\item{aes_y}{A string that sets the y axis variable name from \code{df} and controls the height of
individual \dQuote{sticks}.}

\item{aes_color}{A string that sets the variable name from \code{df} for the aesthetic mapping for color.}

\item{title}{A string that sets the plot title.}

\item{subtitle}{A string that sets the plot subtitle.}

\item{center_titles}{A logical which if \code{TRUE} centers both the \code{title} and \code{subtitle}.}

\item{x_title}{A string that sets the x axis title.}

\item{y_title}{A string that sets the y axis title.}

\item{rot_x_tic_angle}{A numeric that sets the angle of rotation for the x tic labels. When x tic labels are long,
a value of 40 for this argument usually works well.}

\item{rot_y_tic_label}{A logical which if TRUE rotates the y tic labels 90 degrees for enhanced readability.}

\item{x_limits}{Depending on the class of \code{aes_x}, a numeric/Date/POSIXct 2 element vector that sets the minimum
and maximum for the x axis. Use NA to refer to the existing minimum and maximum.}

\item{x_major_breaks}{Depending on the class of \code{aes_x}, a numeric/Date/POSIXct vector or function that
defines the exact major tic locations along the x axis.}

\item{x_labels}{A character vector with the same length as \code{x_major_breaks}, that labels the major tics.}

\item{x_major_date_breaks}{If the class of \code{aes_x} is Date/POSIXct, a string containing the number and date
unit for major breaks. \code{"1 year"}, \code{"4 sec"}, \code{"3 month"}, \code{"2 week"}.}

\item{x_date_labels}{If the class of \code{aes_x} is Date/POSIXct, a string containing the format codes, the
strftime format, for the date. Examples: \code{\%Y-\%m}, \code{\%Y/\%b/\%d}, \code{\%H-\%M-\%S}}

\item{y_limits}{A numeric 2 element vector that sets the minimum and  maximum for the y axis.
Use NA to refer to the existing minimum and maximum.}

\item{y_major_breaks}{A numeric vector or function that defines the exact major tic locations along the y axis.}

\item{y_minor_breaks}{A numeric vector or function that defines the exact minor tic locations along the y axis.}

\item{y_labels}{A character vector with the same length as \code{y_major_breaks}, that labels the major tics.}

\item{axis_text_size}{A numeric that sets the font size along the axis'. Default is 11.}

\item{line_color}{A string that sets the color of the lines.}

\item{line_size}{A numeric value that sets the thickness of lines.}

\item{line_alpha}{A numeric value that sets the degree of color alpha for the lines.}

\item{palette_colors}{A character vector to set the palette colors.}

\item{do_x_title}{A logical that controls the appearence of the x axis title.}

\item{do_y_title}{A logical that controls the appearence of the y axis title.}

\item{show_major_grids}{A logical that controls the appearence of major grids.}

\item{show_minor_grids}{A logical that controls the appearence of minor grids.}

\item{show_legend}{A logical that controls the appearence of the legend.}

\item{legend_pos}{A string that sets the legend position. Acceptable values are
\dQuote{top}, \dQuote{bottom}, \dQuote{left}, \dQuote{right}.}

\item{bold_y}{A numeric that plots a bold horizontal line at this y value.}

\item{silent_NA_warning}{A logical that controls the appearance of a console warning when NA's
are removed.}
}
\value{
A plot object
}
\description{
Function returns a plot object showing vertical/horizontal lines that run from a base value to
a measurement value. Options are provided for scaling.
}
\author{
Rick Dean
}
