% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table.R
\name{create_table}
\alias{create_table}
\title{Function creates a table for tabular data in rmarkdown documents}
\usage{
create_table(
  x = NULL,
  format = "html",
  col_names = NULL,
  font_size = NULL,
  caption = NULL,
  caption_sz = 20,
  align = "c",
  position = "center",
  full_width = F,
  fixed_thead = F,
  head_bkgd = "white",
  head_col = "black",
  head_sz = 18,
  head_angle = 0,
  footnote = NULL,
  footnote_title = NULL,
  scroll_height = NULL
)
}
\arguments{
\item{x}{A data frame or matrix with tabular data.}

\item{format}{A string that sets the tables format. Acceptable values
are \dQuote{html} or \dQuote{pdf}.}

\item{col_names}{A string vector of column names for the table. If NULL
then names are taken from \code{x}.}

\item{font_size}{A numeric that sets the table's font size.}

\item{caption}{A string that sets the table's caption.}

\item{caption_sz}{A numeric that sets the font size of the caption in pixels.}

\item{align}{A string that sets the alignment for the data under the headings.
Acceptable values are \dQuote{c}, \dQuote{l}, \dQuote{r}.}

\item{position}{A string that sets the placement of the table. Acceptable
values are \dQuote{center}, \dQuote{left}, \dQuote{right}.}

\item{full_width}{A logical which if TRUE, the table takes the full width of
the documents available width.}

\item{fixed_thead}{A logical which if TRUE fixes the header as scrolling takes place.}

\item{head_bkgd}{A string that sets the header's background color.}

\item{head_col}{A string that sets the header's color.}

\item{head_sz}{A numeric that sets the header's font size.}

\item{head_angle}{A numeric that sets the header's angle.}

\item{footnote}{A string which if not NULL adds a general footnote to the table.}

\item{footnote_title}{A string that sets the footnote title.}

\item{scroll_height}{A string that sets the scroll height in pixels. For example:
\dQuote{200px}}
}
\value{
A "kableExtra"-"knitr_kable" object
}
\description{
The function is a wrapper around kableExtra::kbl() for
defaulting many of the options available in creating tables in rmarkdown.
A \code{knitr_kable} is returned where further modification can be applied.
}
\author{
Rick Dean
}
