% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_waffle_chart.R
\name{create_waffle_chart}
\alias{create_waffle_chart}
\title{Function creates a "waffle" chart that displays percentages for multiple categorical variables}
\usage{
create_waffle_chart(
  x,
  name_col = NULL,
  prop_col = NULL,
  title = NULL,
  background_color = "yellow",
  cell_fill = "purple",
  cell_color = "black",
  cell_nonfill = "white",
  cell_alpha = 1,
  line_width = 1
)
}
\arguments{
\item{x}{Either a dataframe or a vector of named integer proportions.}

\item{name_col}{If \emph{x} is a dataframe, the column name having the variable names.}

\item{prop_col}{If \emph{x} is a dataframe, the column name having the variable proportions.}

\item{title}{A main title for the chart.}

\item{background_color}{A string that sets the background color.}

\item{cell_fill}{A string that sets the cell fill color.}

\item{cell_color}{A string that sets the cell boundary line color.}

\item{cell_nonfill}{A string that sets the non-cell fill color.}

\item{cell_alpha}{A numeric that sets the cell's fill alpha.}

\item{line_width}{A numeric that sets the line width among the cells}
}
\value{
A ggplot2 plot object
}
\description{
For more information on "waffle" charts see \href{https://baryon.be/uncommon-chart-types-waffle-charts/}{What is a waffle chart}
}
\author{
Rick Dean
}
