% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{farms}
\alias{farms}
\title{farms data}
\format{
Data frame with columns:
\describe{
  \item{state}{character as a state abbreviation}
  \item{count}{numeric of state's total farm count}
}
}
\source{
\href{https://github.com/bayesball/LearnEDAfunctions}{LearnEDAfunctions}
}
\usage{
farms
}
\description{
Number of farms in the states of the U.S. The
raw tab delimited data (\code{farms.txt}) is read by
\code{data.table::fread()} to produce a data frame with columns
for "state" and "count".
}
\keyword{datasets}
