% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gapminder_data}
\alias{gapminder_data}
\title{gapminder_data}
\format{
Data frame with columns:
\describe{
  \item{year}{numeric factor as 1952, 1972, 1992, 2002}
  \item{pop}{population in units of 1e+6}
  \item{continent}{factor with 4 levels}
  \item{lifeExp}{life expectancy at birth, in years}
}
}
\source{
\code{gapminder::gapminder}
}
\usage{
gapminder_data
}
\description{
Based on \code{gapminder::gapminder} dataset, selects 10 year intervals
and 4 continents.  The "year" variable is defined as a factor and the "pop"
variable is divided by 1e+6.
}
\keyword{datasets}
