% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_grob_component.R
\name{get_grob_component}
\alias{get_grob_component}
\title{get_grob_component}
\usage{
get_grob_component(a_plot, component_name)
}
\arguments{
\item{a_plot}{A ggplot2 plot object}

\item{component_name}{A string that sets the components name}
}
\value{
A grob object
}
\description{
Function retrieves a grob component from a ggplot2 plot
object.
}
\author{
Rick Dean
}
