% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kentucky_counties}
\alias{kentucky_counties}
\title{kentucky_counties data}
\format{
Special feature object (class sf) with columns:
\describe{
  \item{ID}{character giving the name of the county}
  \item{geom}{MULTIPOLYGON of the Kentucky county boundary}
}
}
\source{
\code{maps::map()}
}
\usage{
kentucky_counties
}
\description{
A simple feature object with the boundary geometries for nine
southeast Kentucky counties.
.
}
\keyword{datasets}
