% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kentucky_elevation}
\alias{kentucky_elevation}
\title{kentucky_elevation data}
\format{
An object of class \code{data.frame} with 609 rows and 3 columns.
}
\source{
\code{elevatr::get_elev_raster()} and the
Amazon Web Services Terrain Tiles service.
}
\usage{
kentucky_elevation
}
\description{
Tiled raster elevation data(in meters) converted to a data frame,
covering the southeast region of Kentucky.

\tabular{ll}{
  \strong{x} \tab numeric longitude of the elevation \cr
  \strong{y} \tab numeric latitude of the elevation \cr
  \strong{elevation} \tab the elevation in meters
}
}
\keyword{datasets}
