% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{midwest}
\alias{midwest}
\title{midwest data}
\format{
Data frame with columns:
\describe{
  \item{HS_Diploma}{numeric percent high school diploma}
  \item{College_Edu}{numeric percent college educated}
  \item{Prof_Deg}{percentage professional degree}
  \item{White}{percent white}
  \item{Black}{percent black}
  \item{Asian}{percent asian}
}
}
\source{
\code{ggplot2::midwest}
}
\usage{
midwest
}
\description{
From the \code{ggplot2::midwest} dataset, selects for midwest counties
percent high school diploma, percent college educated, percentage professional
degree, percent white, percent black, percent asian.
}
\keyword{datasets}
