% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_bar_plot.R
\name{multi_bar_plot}
\alias{multi_bar_plot}
\title{Produces multiple bar plots.}
\usage{
multi_bar_plot(
  df,
  variables = NULL,
  factor_var = NULL,
  factor_x = NULL,
  columns = 2,
  aes_y = NULL,
  aes_color = NULL,
  aes_fill = NULL,
  position = "stack",
  title = NULL,
  subtitle = NULL,
  x_title = NULL,
  y_titles = NULL,
  rot_x_tic_angle = 0,
  rot_y_tic_label = FALSE,
  bar_fill = NA,
  bar_color = "black",
  bar_alpha = 1,
  bar_size = 1,
  bar_width = NULL,
  palette_colors = NULL,
  x_limits = NULL,
  x_breaks = waiver(),
  x_labels = waiver(),
  x_date_breaks = waiver(),
  x_date_labels = waiver(),
  y_limits = NULL,
  y_major_breaks = waiver(),
  y_minor_breaks = waiver(),
  y_labels = waiver(),
  axis_text_size = 11,
  col_width = 4,
  row_height = 4,
  do_coord_flip = FALSE,
  order_bars = NULL,
  bar_labels = FALSE,
  bar_label_size = 6,
  bar_label_color = "black",
  add_ons = NULL,
  show_major_grids = TRUE,
  show_minor_grids = TRUE,
  show_legend = FALSE,
  legend_pos = "top",
  silent_NA_warning = FALSE,
  display_plot = TRUE
)
}
\arguments{
\item{df}{The source data frame for the bar chart.}

\item{variables}{A vector that names the variables from \code{df} for plotting bar charts.}

\item{factor_var}{A string that sets the factor variable from \code{df} with multiple levels.}

\item{factor_x}{A string that sets a variable name from \code{df}.  If \code{factor_var} is not NULL
then a bar chart of \code{factor_x} will be plotted for each of \code{factor_var}'s levels.}

\item{columns}{An integer the sets the number of columns for the multi-panel display.}

\item{aes_y}{A string that sets the y axis variable name from \code{df}.}

\item{aes_color}{A string that sets the variable name from \code{df} for the aesthetic mapping for color.}

\item{aes_fill}{A string that sets the variable name from \code{df} for the aesthetic mapping for fill.}

\item{position}{A string that sets the bar positions. Acceptable values are \code{dodge}(side by side),
\code{identity}(overlap) or \code{stack}.}

\item{title}{A string that sets the overall title.}

\item{subtitle}{A string that sets the overall subtitle.}

\item{x_title}{A string that sets the x axis title.}

\item{y_titles}{A string vector that sets the y axis title corresponding to each variable in \code{variables} or
each factor level in \code{factor_var}. If \code{y_titles} is \code{NULL} then the y axis title is \dQuote{Count}.}

\item{rot_x_tic_angle}{A numeric that sets the angle of rotation for the x tic labels. When x tic labels are long,
a value of 40 for this argument usually works well.}

\item{rot_y_tic_label}{A logical which if TRUE rotates the y tic labels 90 degrees for enhanced readability.}

\item{bar_fill}{A string that sets the fill color for the bars.}

\item{bar_color}{A string that sets the outline color for the bars.}

\item{bar_alpha}{A numeric that sets the alpha component to \code{bar_color}.}

\item{bar_size}{A numeric that sets the outline thickness of the bars.}

\item{bar_width}{A numeric that sets the width of the bars.}

\item{palette_colors}{A character vector to set the palette colors.}

\item{x_limits}{A character vector that defines possible values of the scale and their order..}

\item{x_breaks}{A character vector of x axis breaks or a function that takes the limits as input and returns breaks
as output.}

\item{x_labels}{A character vector that defines the x axis tic labels. Vector must be the same length as
\code{x_breaks}.}

\item{x_date_breaks}{If the class of \code{variables} or \code{factor_x} is Date/POSIXct, a string containing the number and date unit for
major breaks. \code{"1 years"}, \code{"4 sec"}, \code{"3 month"}, \code{"2 week"}.}

\item{x_date_labels}{If the class of \code{variables} or \code{factor_x} is Date/POSIXct, a string containing the format codes, the
strftime format, for the date. Examples: \code{\%Y-\%m}, \code{\%Y/\%b/\%d}, \code{\%H-\%M-\%S}}

\item{y_limits}{A numeric 2 element vector that sets the minimum and  maximum for the y axis.}

\item{y_major_breaks}{A numeric vector or function that defines the exact major tic locations along the y axis.}

\item{y_minor_breaks}{A numeric vector or function that defines the exact minor tic locations along the y axis.}

\item{y_labels}{A character vector with the same length as \code{y_major_breaks}, that labels the major tics.}

\item{axis_text_size}{A numeric that sets the font size along the axis'. Default is 11.}

\item{col_width}{An integer that sets the width of each plot column in inches.}

\item{row_height}{An integer that sets the height of each plot column in inches.}

\item{do_coord_flip}{A logical which if \code{TRUE} will flip the x and y axis'.}

\item{order_bars}{A string which will order the bars in a specific direction. Acceptable values are \dQuote{asc} or \dQuote{desc}}

\item{bar_labels}{A logical which if \code{TRUE} will label each bar with its value.}

\item{bar_label_size}{A numeric that sets the size of the bar labels}

\item{bar_label_color}{A string that sets the color of the bar labels}

\item{add_ons}{Is a vector of ggplot objects that will be appended to each bar plot object.  This can be used
to override or add to all the multiple bar plot objects.}

\item{show_major_grids}{A logical that controls the appearance of major grids.}

\item{show_minor_grids}{A logical that controls the appearance of minor grids.}

\item{show_legend}{A logical that controls the appearance of the legend.}

\item{legend_pos}{A string that sets the legend position. Acceptable values are
\dQuote{top}, \dQuote{bottom}, \dQuote{left}, \dQuote{right}.}

\item{silent_NA_warning}{A logical that controls the appearance of a console warning when Na's
are removed.}

\item{display_plot}{A logical that if TRUE will display the plot}
}
\value{
A TableGrob object.
}
\description{
The function creates multiple bar plots from one of two sources:
\enumerate{
\item From a vector containing selected names of a data frame's variables.
\item A single variable's bar plot mapped across another variable's factor levels.
}

The function returns a single multi-paneled plot object for display.
}
\author{
Rick Dean
}
