% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_density_plot.R
\name{multi_density_plot}
\alias{multi_density_plot}
\title{Produces multiple ggplot2 based density plots.}
\usage{
multi_density_plot(
  df,
  kernel = "gaussian",
  bw = "nrd0",
  variables = NULL,
  factor_var = NULL,
  factor_x = NULL,
  columns = 2,
  aes_color = NULL,
  aes_fill = NULL,
  position = "stack",
  title = NULL,
  subtitle = NULL,
  x_title = NULL,
  y_title = NULL,
  y_titles = NULL,
  rot_x_tic_angle = 0,
  rot_y_tic_label = FALSE,
  density_size = 1,
  density_fill = NA,
  density_color = "black",
  density_alpha = 1,
  palette_colors = NULL,
  x_limits = NULL,
  x_major_breaks = waiver(),
  x_minor_breaks = waiver(),
  x_labels = waiver(),
  y_limits = NULL,
  y_major_breaks = waiver(),
  y_minor_breaks = waiver(),
  y_labels = waiver(),
  axis_text_size = 11,
  col_width = 4,
  row_height = 4,
  add_ons = NULL,
  show_major_grids = TRUE,
  show_minor_grids = TRUE,
  plot_obs = FALSE,
  plot_obs_len = 0.02,
  plot_obs_color = "black",
  plot_obs_jitter = FALSE,
  show_legend = FALSE,
  legend_pos = "top",
  na.rm = TRUE,
  display_plot = TRUE
)
}
\arguments{
\item{df}{The target data frame for the density chart.}

\item{kernel}{A string that set the type of Kernel Density Estimation (KDE). Acceptable values are "gaussian",
"rectangular", "triangular", "epanechnikov", "biweight", "cosine" or "optcosine".}

\item{bw}{A string or numeric that sets the smoothing bandwidth to be used with the KDE function.}

\item{variables}{A vector that names the variables from \code{df} for plotting densities.}

\item{factor_var}{A string that sets the factor variable from \code{df} with multiple levels.}

\item{factor_x}{A string that sets a variable name from \code{df}.  If \code{factor_var} is not NULL
then a density distribution of \code{factor_x} will be plotted for each of \code{factor_var}'s levels.}

\item{columns}{An integer the sets the number of columns for the multi-panel display.}

\item{aes_color}{A string that sets the variable name from \code{df} for the aesthetic mapping for color.}

\item{aes_fill}{A string that sets the variable name from \code{df} for the aesthetic mapping for fill.}

\item{position}{A string that sets the position.  Acceptable values are \code{identity} which overlays or
\code{stack} which stacks.}

\item{title}{A string that sets the overall title.}

\item{subtitle}{A string that sets the overall subtitle.}

\item{x_title}{A string that sets the x axis title.}

\item{y_title}{A string that sets the y axis title for all variables in \code{variables}.}

\item{y_titles}{A string vector that sets the y axis title corresponding to each variable in \code{variables} or
each factor level in \code{factor_var}. If \code{y_titles} is \code{NULL} then the y axis title is \dQuote{Density}.}

\item{rot_x_tic_angle}{A numeric that sets the angle of rotation for the x tic labels. When x tic labels are long,
a value of 40 for this argument usually works well.}

\item{rot_y_tic_label}{A logical which if TRUE rotates the y tic labels 90 degrees for enhanced readability.}

\item{density_size}{A numeric that sets the density line width.}

\item{density_fill}{A string that sets the fill color for the density curve.}

\item{density_color}{A string that sets the color for the density line.}

\item{density_alpha}{A numeric that set the alpha component to \code{density_color}.}

\item{palette_colors}{A character vector to set the palette colors.}

\item{x_limits}{A numeric 2 element vector or function that sets the minimum and maximum for the x axis.  Use NA to
refer to the existing minimum and maximum.}

\item{x_major_breaks}{A numeric vector or function that sets the major tic locations along the x axis.}

\item{x_minor_breaks}{A numeric vector or function that sets the minor tic locations along the x axis.}

\item{x_labels}{A character vector or function giving x axis tic labels.  Must be the same length as \code{x_breaks}.}

\item{y_limits}{A numeric 2 element vector or function that sets the minimum and maximum for the y axis.  Use NA to
refer to the existing minimum and maximum.}

\item{y_major_breaks}{A numeric vector or function that sets the major tic locations along the y axis.}

\item{y_minor_breaks}{A numeric vector or function that sets the minor tic locations along the y axis.}

\item{y_labels}{A character vector or function giving y axis tic labels.  Must be the same length as \code{y_breaks}.}

\item{axis_text_size}{A numeric that sets the font size along the axis'. Default is 11.}

\item{col_width}{An integer that sets the width of each plot column in inches.}

\item{row_height}{An integer that sets the height of each plot column in inches.}

\item{add_ons}{Is a vector of ggplot objects that will be appended to each density plot object.  This can be used
to override or add to all the multiple density plot objects.}

\item{show_major_grids}{A logical that controls the appearance of major grids.}

\item{show_minor_grids}{A logical that controls the appearance of minor grids.}

\item{plot_obs}{A logical which if \code{TRUE} plots a line for each observation along the axis margin.}

\item{plot_obs_len}{A numeric that sets the length of the \code{plot_obs} lines.}

\item{plot_obs_color}{A string that sets the color of the \code{plot_obs} lines.}

\item{plot_obs_jitter}{A logical which if \code{TRUE} will add a slight horizontal adjustment to overlapping observations.}

\item{show_legend}{A logical that controls the appearance of the legend.}

\item{legend_pos}{A string that sets the legend position. Acceptable values are
\dQuote{top}, \dQuote{bottom}, \dQuote{left}, \dQuote{right}.}

\item{na.rm}{A logical which if \code{TRUE}, missing values are removed from \code{df}. If \code{FALSE} any missing
values cause an error.}

\item{display_plot}{A logical that if TRUE will display the plot.}
}
\value{
A TableGrob object.
}
\description{
The function creates multiple density plots from one of two sources:
\enumerate{
\item From a vector containing selected names of a data frame's variables.
\item A single variable's densities mapped across levels of a data frame's factor variable.
}

The function returns a single multi-paneled plot object for display.
}
\author{
Rick Dean
}
