% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_range_plot.R
\name{multi_range_plot}
\alias{multi_range_plot}
\title{Produces multiple range plots.}
\usage{
multi_range_plot(
  df,
  factor_var = NULL,
  factor_x = NULL,
  columns = 2,
  col_width = 4,
  row_height = 4,
  aes_y = NULL,
  aes_y_min,
  aes_y_max,
  title = NULL,
  subtitle = NULL,
  center_titles = FALSE,
  x_title = NULL,
  y_titles = NULL,
  rot_x_tic_angle = 0,
  rot_y_tic_label = FALSE,
  y_limits = NULL,
  y_major_breaks = waiver(),
  y_minor_breaks = waiver(),
  y_labels = waiver(),
  y_log10 = FALSE,
  axis_text_size = 11,
  pts_fill = "white",
  pts_shape = 21,
  pts_stroke = 1,
  line_type = "solid",
  line_pts_size = 1,
  line_pts_color = "black",
  line_pts_alpha = 1,
  do_x_title = TRUE,
  do_y_title = TRUE,
  show_major_grids = TRUE,
  show_minor_grids = TRUE,
  display_plot = TRUE,
  do_coord_flip = FALSE,
  add_ons = NULL,
  silent_NA_warning = FALSE
)
}
\arguments{
\item{df}{The source data frame from which the x/y pairs and factor variable originate.}

\item{factor_var}{A string that sets the factor variable from \code{df} whose levels will be plotted.}

\item{factor_x}{A string that sets a variable name from \code{df}. A range plot of
\code{factor_x} versus \code{aes_y} will be plotted for each of \code{factor_var}'s levels.}

\item{columns}{An integer that sets the number of columns for the multi-panel display.}

\item{col_width}{An integer that sets the width of each plot column in inches.}

\item{row_height}{An integer that sets the height of each plot column in inches.}

\item{aes_y}{A string that sets the y axis variable name from \code{df}.}

\item{aes_y_min}{A string that sets a y axis variable name from \code{df}.
These are the required numerics that defines the minimum values for the range
of \code{aes_y}.}

\item{aes_y_max}{A string that sets a y axis variable name from \code{df}.
These are the required numerics that defines the maximum values for the range
of \code{aes_y}.}

\item{title}{A string that sets the plot title.}

\item{subtitle}{A string that sets the plot subtitle.}

\item{center_titles}{A logical which if \code{TRUE} centers both the \code{title}
and \code{subtitle}.}

\item{x_title}{A string that sets the x axis title.}

\item{y_titles}{A string vector that sets the y axis title corresponding to each factor level. If \code{y_titles}
is \code{NULL} then the y axis title is the same as \code{aes_y}.}

\item{rot_x_tic_angle}{A numeric that sets the angle of rotation for the x tic labels.
When x tic labels are long, a value of 40 for this argument usually works well.}

\item{rot_y_tic_label}{A logical which if TRUE rotates the y tic labels 90 degrees
for enhanced readability.}

\item{y_limits}{A numeric 2 element vector that sets the minimum and  maximum for the y axis.
Use \code{NA} to refer to the existing minimum and maximum.}

\item{y_major_breaks}{A numeric vector or function that defines the exact major tic locations along the y axis.}

\item{y_minor_breaks}{A numeric vector or function that defines the exact minor tic locations along the y axis.}

\item{y_labels}{A character vector with the same length as \code{y_major_breaks}, that labels the major tics.}

\item{y_log10}{A logical which if \code{TRUE} will use a log10 scale for the y axis.}

\item{axis_text_size}{A numeric that sets the font size along the axis'. Default is 11.}

\item{pts_fill}{A string that sets the fill color of the points.}

\item{pts_shape}{A numeric integer that sets the shape of the points. Typical values are 21 \dQuote{circle},
22 \dQuote{square}, 23 \dQuote{diamond}, 24 \dQuote{up triangle}, 25 \dQuote{down triangle}.}

\item{pts_stroke}{A numeric that sets the drawing width for a point shape.}

\item{line_type}{A string that sets range line type \code{twodash, solid, longdash, dotted, dotdash,
dashed, blank}.}

\item{line_pts_size}{A numeric value that sets the size of both lines(width) and points(diameter).}

\item{line_pts_color}{A string that sets the color of the range lines and outlines of the points.}

\item{line_pts_alpha}{A numeric value that sets the alpha level of points.}

\item{do_x_title}{A logical that controls the appearance of the x axis title.}

\item{do_y_title}{A logical that controls the appearance of the y axis title.}

\item{show_major_grids}{A logical that controls the appearance of major grids.}

\item{show_minor_grids}{A logical that controls the appearance of minor grids.}

\item{display_plot}{A logical that if TRUE will display the plot}

\item{do_coord_flip}{A logical which if \code{TRUE} will flip the x and y axis'.}

\item{add_ons}{Is a vector of ggplot objects that will be appended to separate scatter plots. Each
element in this vector addresses a separate plot in the order of presentation.}

\item{silent_NA_warning}{A logical that controls the appearance of a console warning when Na's
are removed.}
}
\value{
A TableGrob object.
}
\description{
The function creates multiple range plots from x/y pairs across
the levels of a separate selected factor variable.
}
\author{
Rick Dean
}
