% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_scatter_plot.R
\name{multi_scatter_plot}
\alias{multi_scatter_plot}
\title{Produces multiple scatter plots.}
\usage{
multi_scatter_plot(
  df,
  variables = NULL,
  factor_var = NULL,
  factor_x = NULL,
  columns = 2,
  aes_y = NULL,
  aes_color = NULL,
  aes_fill = NULL,
  aes_size = NULL,
  aes_linetype = NULL,
  aes_label = NULL,
  position = ggplot2::position_jitter(width = 0, height = 0),
  title = NULL,
  subtitle = NULL,
  x_title = NULL,
  y_titles = NULL,
  rot_x_tic_angle = 0,
  rot_y_tic_label = FALSE,
  x_limits = NULL,
  x_major_breaks = waiver(),
  x_minor_breaks = waiver(),
  x_labels = waiver(),
  x_major_date_breaks = waiver(),
  x_minor_date_breaks = waiver(),
  x_date_labels = waiver(),
  x_log10 = FALSE,
  y_limits = NULL,
  y_major_breaks = waiver(),
  y_minor_breaks = waiver(),
  y_labels = waiver(),
  axis_text_size = 11,
  show_pts = TRUE,
  y_log10 = FALSE,
  pts_fill = "white",
  pts_shape = 21,
  pts_stroke = 1,
  pts_color = "black",
  pts_line_alpha = 1,
  pts_size = 1,
  palette_colors = NULL,
  palette_linetypes = NULL,
  connect = FALSE,
  line_size = 0.6,
  line_color = "black",
  connect_linetype = "solid",
  col_width = 4,
  row_height = 4,
  show_major_grids = TRUE,
  show_minor_grids = TRUE,
  show_legend = FALSE,
  legend_pos = "top",
  display_plot = TRUE,
  bold_y = NULL,
  CI_lwr = NULL,
  CI_upr = NULL,
  CI_dir = "y",
  CI_type = "bar",
  CI_color = "black",
  CI_width = 0.1,
  CI_size = 1,
  add_ons = NULL,
  silent_NA_warning = FALSE
)
}
\arguments{
\item{df}{The source data frame from which the scatter points are plotted.}

\item{variables}{A vector that names the x axis variables from \code{df} for plotting these variables
versus \code{aes_y} in multiple scatter plots.}

\item{factor_var}{A string that sets the factor variable from \code{df} whose levels will be plotted.}

\item{factor_x}{A string that sets a variable name from \code{df}.  If \code{factor_var} is not NULL
then a scatter plot of \code{factor_x} versus \code{aes_y} will be plotted for each of \code{factor_var}'s levels.}

\item{columns}{An integer that sets the number of columns for the multi-panel display.}

\item{aes_y}{A string that sets the y axis variable name from \code{df}.}

\item{aes_color}{A string that sets the variable name from \code{df} for the aesthetic mapping for color.}

\item{aes_fill}{A string that sets the variable name from \code{df} for the aesthetic mapping for fill.}

\item{aes_size}{A string that sets the variable name from \code{df} for the aesthetic mapping for size.}

\item{aes_linetype}{A string that sets the variable name from \code{df} for the aesthetic mapping for linetype.}

\item{aes_label}{A string that sets the variable name from \code{df} for the aesthetic mapping for labeling. If
labelling points then the package \code{ggrepel} is required.}

\item{position}{A string or function that does a slight adjustment to overlapping points.  Typical values are
\code{jitter} or \code{position_jitter(width = 0.1, height = 0.1)}.}

\item{title}{A string that sets the plot title.}

\item{subtitle}{A string that sets the plot subtitle.}

\item{x_title}{A string that sets the x axis title.}

\item{y_titles}{A string vector that sets the y axis title corresponding to each factor level. If \code{y_titles}
is \code{NULL} then the y axis title is the same as \code{aes_y}.}

\item{rot_x_tic_angle}{A numeric that sets the angle of rotation for the x tic labels. When x tic labels are long,
a value of 40 for this argument usually works well.}

\item{rot_y_tic_label}{A logical which if TRUE rotates the y tic labels 90 degrees for enhanced readability.}

\item{x_limits}{Depending on the class of \code{aes_x}, a numeric/Date/POSIXct 2 element vector that sets the
minimum and maximum for the x axis.}

\item{x_major_breaks}{Depending on the class of \code{aes_x}, a numeric/Date/POSIXct vector or function that defines
the exact major tic locations along the x axis.}

\item{x_minor_breaks}{Depending on the class of \code{aes_x}, a numeric/Date/POSIXct vector or function that defines
the exact minor tic locations along the x axis.}

\item{x_labels}{A character vector with the same length as \code{x_major_breaks}, that labels the major tics.}

\item{x_major_date_breaks}{If the class of \code{aes_x} is Date/POSIXct, a string containing the number and date unit
for major breaks. Examples: \code{"1 year"}, \code{"4 sec"}, \code{"3 month"}, \code{"2 week"}.}

\item{x_minor_date_breaks}{If the class of \code{aes_x} is Date/POSIXct, a string containing the number and date unit
for minor breaks.}

\item{x_date_labels}{If the class of \code{aes_x} is Date/POSIXct, a string containing the format codes, the strftime
format, for the date. Examples: \code{"\%Y-\%m"}, \code{"\%Y/\%b/\%d"}, \code{"\%H-\%M-\%S"}.}

\item{x_log10}{A logical which if \code{TRUE} will use a log10 scale for the x axis.}

\item{y_limits}{A numeric 2 element vector that sets the minimum and  maximum for the y axis.}

\item{y_major_breaks}{A numeric vector or function that defines the exact major tic locations along the y axis.}

\item{y_minor_breaks}{A numeric vector or function that defines the exact minor tic locations along the y axis.}

\item{y_labels}{A character vector with the same length as \code{y_major_breaks}, that labels the major tics.}

\item{axis_text_size}{A numeric that sets the font size along the axis'. Default is 11.}

\item{show_pts}{A logical which if FALSE will plot only the lines if \code{connect} is TRUE.}

\item{y_log10}{A logical which if TRUE will use a log10 scale for the y axis.}

\item{pts_fill}{A string that sets the fill color of the points.}

\item{pts_shape}{A numeric integer that sets the shape of the points. Typical values are 21 \dQuote{circle},
22 \dQuote{square}, 23 \dQuote{diamond}, 24 \dQuote{up triangle}, 25 \dQuote{down triangle}.}

\item{pts_stroke}{A numeric that sets the drawing width for a point shape.}

\item{pts_color}{A string that sets the color of the points.}

\item{pts_line_alpha}{A numeric value that sets the alpha level of \code{pts_line_color}.}

\item{pts_size}{A numeric value that sets the size of the points.}

\item{palette_colors}{A string vector to set the aes mapping of colors which changes their appearance in the
order given.}

\item{palette_linetypes}{A string vector to set the aes mapping of linetypes which changes their appearance in
the order given.}

\item{connect}{A logical which if \code{TRUE} then points will be connected with a line.}

\item{line_size}{A numeric value that sets the thickness of lines if \code{connect} is TRUE.}

\item{line_color}{A string that sets the color of the lines if \code{connect} is TRUE.}

\item{connect_linetype}{A string that sets line type \code{twodash, solid, longdash, dotted, dotdash,
dashed, blank} if \code{connect} is TRUE.}

\item{col_width}{An integer that sets the width of each plot column in inches.}

\item{row_height}{An integer that sets the height of each plot column in inches.}

\item{show_major_grids}{A logical that controls the appearance of major grids.}

\item{show_minor_grids}{A logical that controls the appearance of minor grids.}

\item{show_legend}{A logical that controls the appearance of the legend.}

\item{legend_pos}{A string that sets the legend position. Acceptable values are
\dQuote{top}, \dQuote{bottom}, \dQuote{left}, \dQuote{right}.}

\item{display_plot}{A logical that if TRUE will display the plot}

\item{bold_y}{A numeric that plots a bold horizontal line at this y value intercept.}

\item{CI_lwr}{A string that sets the column from \code{df} for the lower confidence interval with reference to the x or y axis.}

\item{CI_upr}{A string that sets the column from \code{df} for the upper confidence interval with reference to the x or y axis.}

\item{CI_dir}{A string that sets the axis orientation of the confidence intervals. Acceptable values are "x" or "y".}

\item{CI_type}{A string that sets the confidence interval type. Acceptable values are "bar" or "ribbon".}

\item{CI_color}{A string that sets the confidence interval color.}

\item{CI_width}{A numeric that sets the confidence interval error cross bar width.}

\item{CI_size}{A numeric that sets the confidence interval line thickness.}

\item{add_ons}{Is a vector of ggplot objects that will be appended to separate scatter plots. Each
element in this vector addresses a separate plot in the order of presentation.}

\item{silent_NA_warning}{A logical that controls the appearance of a console warning when NA's
are removed at the start/end of the line.}
}
\value{
A TableGrob object.
}
\description{
The function creates multiple scatter plots from one of two sources:
\enumerate{
\item From a vector containing selected names of a data frame's variables.  These x axis variables
are paired with a single y axis variable in multiple panels of x versus y scatter plots.
\item A single variable's scatter plot mapped across levels of a data frame's factor variable.
}

The function returns a single multi-paneled plot object for display.
}
\author{
Rick Dean
}
