% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_stick_plot.R
\name{multi_stick_plot}
\alias{multi_stick_plot}
\title{Produces multiple stick plots.}
\usage{
multi_stick_plot(
  df,
  variables = NULL,
  factor_var = NULL,
  factor_x = NULL,
  columns = 2,
  base_val = 0,
  aes_y = NULL,
  aes_color = NULL,
  title = NULL,
  subtitle = NULL,
  x_title = NULL,
  y_titles = NULL,
  rot_x_tic_angle = 0,
  rot_y_tic_label = FALSE,
  x_limits = NULL,
  x_major_breaks = waiver(),
  x_labels = waiver(),
  x_major_date_breaks = waiver(),
  x_date_labels = waiver(),
  y_limits = NULL,
  y_major_breaks = waiver(),
  y_labels = waiver(),
  axis_text_size = 11,
  col_width = 4,
  row_height = 4,
  line_color = "black",
  line_size = 0.8,
  palette_colors = NULL,
  add_ons = NULL,
  show_major_grids = TRUE,
  show_minor_grids = TRUE,
  show_legend = FALSE,
  legend_pos = "top",
  display_plot = TRUE,
  bold_y = NULL,
  silent_NA_warning = FALSE
)
}
\arguments{
\item{df}{The source data frame from which the sticks are plotted.}

\item{variables}{A vector that names the variables from \code{df} for the x axis.}

\item{factor_var}{A string that sets the factor variable from \code{df} with multiple levels.}

\item{factor_x}{A string that sets a variable name from \code{df}.  If \code{factor_var} is not NULL
then a stick plot of \code{factor_x} will be plotted for each of \code{factor_var}'s levels.}

\item{columns}{An integer the sets the number of columns for the multi-panel display.}

\item{base_val}{A numeric that sets the base value from which the \dQuote{stick} originates.
The default value is 0.}

\item{aes_y}{A string that sets the y axis variable name from \code{df} and controls the height of
individual \dQuote{sticks}.}

\item{aes_color}{A string that sets the variable name from \code{df} for the aesthetic mapping for color.}

\item{title}{A string that sets the plot title.}

\item{subtitle}{A string that sets the plot subtitle.}

\item{x_title}{A string that sets the x axis title.}

\item{y_titles}{A string vector that sets the y axis title corresponding to each factor level. If \code{y_titles}
is \code{NULL} then the y axis title is the same as \code{aes_y}.}

\item{rot_x_tic_angle}{A numeric that sets the angle of rotation for the x tic labels. When x tic labels are long,
a value of 40 for this argument usually works well.}

\item{rot_y_tic_label}{A logical which if TRUE rotates the y tic labels 90 degrees for enhanced readability.}

\item{x_limits}{Depending on the class of \code{variables}, a numeric/Date/POSIXct 2 element vector that sets the minimum
and maximum for the x axis. Use NA to refer to the existing minimum and maximum.}

\item{x_major_breaks}{Depending on the class of \code{variables}, a numeric/Date/POSIXct vector or function that
defines the exact major tic locations along the x axis.}

\item{x_labels}{A character vector with the same length as \code{x_major_breaks}, that labels the major tics.}

\item{x_major_date_breaks}{If the class of \code{variables} is Date/POSIXct, a string containing the number and date
unit for major breaks. \code{"1 year"}, \code{"4 sec"}, \code{"3 month"}, \code{"2 week"}.}

\item{x_date_labels}{If the class of \code{variables} is Date/POSIXct, a string containing the format codes, the
strftime format, for the date. Examples: \code{\%Y-\%m}, \code{\%Y/\%b/\%d}, \code{\%H-\%M-\%S}}

\item{y_limits}{A numeric 2 element vector that sets the minimum and  maximum for the y axis.
Use NA to refer to the existing minimum and maximum.}

\item{y_major_breaks}{A numeric vector or function that defines the exact major tic locations along the y axis.}

\item{y_labels}{A character vector with the same length as \code{y_major_breaks}, that labels the major tics.}

\item{axis_text_size}{A numeric that sets the font size along the axis'. Default is 11.}

\item{col_width}{An integer that sets the width of each plot column in inches.}

\item{row_height}{An integer that sets the height of each plot column in inches.}

\item{line_color}{A string that sets the color of the lines.}

\item{line_size}{A numeric value that sets the thickness of lines.}

\item{palette_colors}{A character vector to set the palette colors.}

\item{add_ons}{Is a vector of ggplot objects that will be appended to each stick plot object.  This can be used
to override or add to all the multiple stick plot objects.}

\item{show_major_grids}{A logical that controls the appearance of major grids.}

\item{show_minor_grids}{A logical that controls the appearance of minor grids.}

\item{show_legend}{A logical that controls the appearance of the legend.}

\item{legend_pos}{A string that sets the legend position. Acceptable values are
\dQuote{top}, \dQuote{bottom}, \dQuote{left}, \dQuote{right}.}

\item{display_plot}{A logical that if TRUE will display the plot}

\item{bold_y}{A numeric that plots a bold horizontal line at this y value.}

\item{silent_NA_warning}{A logical that controls the appearance of a console warning when Na's
are removed.}
}
\value{
A TableGrob object.
}
\description{
The function creates multiple histogram plots from one of two sources:
\enumerate{
\item From a vector containing selected names of a data frame's variables.
\item A single variable's histograms mapped across levels of a data frame's factor variable.
}

The function returns a single multi-paneled plot object for display.
}
\author{
Rick Dean
}
