% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{organdata}
\alias{organdata}
\title{organdata data}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 204 rows and 2 columns.
}
\source{
\code{socviz::organdata}
}
\usage{
organdata
}
\description{
From the \code{socviz::organdata} dataset two variables are
selected: Organ Donation rate per million population and
country.

\tabular{ll}{
  \strong{country} \tab character country name \cr
  \strong{donors} \tab Organ Donation rate per million population
}
}
\keyword{datasets}
