% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{penguins_stats}
\alias{penguins_stats}
\title{penguins_stats data}
\format{
Data frame with columns:
\describe{
  \item{avg_body_mass}{numeric species average body mass}
  \item{min_body_mass}{numeric species minimum body mass}
  \item{max_body_mass}{numeric species maximum body mass}
}
}
\source{
\code{palmerpenguins:penguins}
}
\usage{
penguins_stats
}
\description{
The \code{palmerpenguins:penguins} dataset is grouped by the variable
"species" and a data frame is produced with statistics for
average, minimum and maximum body mass.
}
\keyword{datasets}
