% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{people}
\alias{people}
\title{people data}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 4 rows and 6 columns.
}
\source{
R coding with \code{data.table::data.table()}
}
\usage{
people
}
\description{
A made-up data frame with generic column names and mix of
numeric and character data.

\tabular{ll}{
  \strong{col_0} \tab numeric random numbers for ids \cr
  \strong{col_1} \tab character first names \cr
  \strong{col_2} \tab character last names \cr
  \strong{col_3} \tab numeric ages \cr
  \strong{col_4} \tab character addresses \cr
  \strong{col_5} \tab character spouse's first name
}
}
\keyword{datasets}
