% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentile_table.R
\name{percentile_table}
\alias{percentile_table}
\title{Function locates values corresponding to  percentiles in a vector of numeric data}
\usage{
percentile_table(vals, na_rm = TRUE, title = NULL, table_width = NULL)
}
\arguments{
\item{vals}{A vector of numeric values}

\item{na_rm}{A logical which if TRUE removes NA values before
the calculations}

\item{title}{A string that sets the table's title. The string can contain R markdown/html style syntax.}

\item{table_width}{A numeric that sets the overall table width in pixels.}
}
\value{
A named list with a gt object and a data.table with percentile values contained in the display table.
}
\description{
Function displays a table with lower/upper locations, values,
average and spread between lower/upper from median to one hundred and twenty-eighth percentiles.
The function returns a list with a viewable gt table along with the data.table
containing the table's values.
}
\details{
The "label" column of the table correspond to the following percentiles:

\enumerate{
\item "M" = Half (i.e. the median)
\item "F" = Fourth (i.e. quartiles 25/75)
\item "E" = Eighth
\item "D" = Sixteenth
\item "C" = Thirty-secondth
\item "B" = Sixty-fourth
\item "A" = One hundred and twenty-eighth
}
}
\author{
Rick Dean
}
