% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{religion}
\alias{religion}
\title{religion data}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 2844 rows and 2 columns.
}
\source{
\code{socviz::gss_sm}
}
\usage{
religion
}
\description{
From the \code{socviz::gss_sm} dataset General Social Survey, 2016,
a sub-data frame was created from the "happy" and "religion" variables.

\tabular{ll}{
  \strong{happy} \tab general happiness \cr
  \strong{religion} \tab religion in six categories
}
}
\keyword{datasets}
