% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sleep}
\alias{sleep}
\title{sleep data}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 20 rows and 4 columns.
}
\source{
\code{datasets::sleep}
}
\usage{
sleep
}
\description{
A dataframe with 20 observations that shows effect of two soporific
drugs (increase in hours of sleep compared to control) on 10 patients.

\tabular{ll}{
  \strong{Hours} \tab extra numeric increase in hours of sleep \cr
  \strong{Drug} \tab factor with 2 levels of drug groups (1,2) \cr
  \strong{ID} \tab factor with 10 levels of patient ID (1 to 10) \cr
  \strong{Observation} \tab numeric observation id (1 to 20)
}
}
\keyword{datasets}
