% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spinrates}
\alias{spinrates}
\title{spinrates data}
\format{
Data frame with columns:
\describe{
  \item{id}{character giving the observation id}
  \item{velocity}{numeric average velocity of the ball}
  \item{spinrate}{character factor with 12 levels}
  \item{swing_miss}{numeric rate of swing and miss}
}
}
\source{
\href{https://www.fangraphs.com}{Fangraphs}
}
\usage{
spinrates
}
\description{
From the raw data file \code{spinrates.csv} the comma delimited data
was converted to a data frame using \code{data.table::fread()}.
The data are baseball percent "swing and miss" with their respective average
ball velocity and levels of spinrate.
}
\keyword{datasets}
