% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{states}
\alias{states}
\title{states data}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 16 rows and 3 columns.
}
\source{
\code{datasets::state_name, state_region, state_x77}
}
\usage{
states
}
\description{
Consolidation of data into one dataframe for the "South" region.

\tabular{ll}{
  \strong{State} \tab character state name \cr
  \strong{Income} \tab numeric state income \cr
  \strong{Murder} \tab numeric state murder rate
}
}
\keyword{datasets}
