% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{world_coffee}
\alias{world_coffee}
\title{world_coffee data}
\format{
Special feature object (class sf) with columns:
\describe{
  \item{name_long}{character long name of country}
  \item{coffee_production_2017}{numeric of coffee production of 60 kg bags}
  \item{geom}{the sfc_MULTIPOLYGON country geometries}
}
}
\source{
\code{spData::coffee_data} and \code{spData::world}
}
\usage{
world_coffee
}
\description{
World coffee production by country (in thousands of 60 kg bags)
  is joined with a simple feature of country geometries via their
  common parameter `name_long`.
}
\keyword{datasets}
