% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{air_passengers}
\alias{air_passengers}
\title{air_passengers data}
\format{
Data frame with columns:
\describe{
  \item{time}{Date with format 1949-01-01}
  \item{value}{numeric number of passengers in thousands}
}
}
\source{
\code{datasets::AirPassengers}
}
\usage{
air_passengers
}
\description{
Data where the base \code{datasets::AirPassengers} monthly time
  series of international airline passenger numbers 1949-1960 was converted to
  a data frame using \code{tsbox::ts_df(datasets::AirPassengers)}.
}
\keyword{datasets}
