% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boston_marathon}
\alias{boston_marathon}
\title{boston_marathon data}
\format{
Data frame with columns:
\describe{
  \item{age}{numeric age from 20 to 60}
  \item{time}{numeric marathon time}
}
}
\source{
\href{https://github.com/bayesball/LearnEDAfunctions}{LearnEDAfunctions}
}
\usage{
boston_marathon
}
\description{
Boston Marathon completion times of women of different ages. The
raw tab delimited data (\code{boston_marathon.txt}) is read by
\code{data.table::fread()} to produce a data frame with columns
for women's "age" and marathon "time" in minutes.
}
\keyword{datasets}
