% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{car_stats}
\alias{car_stats}
\title{car_stats data}
\format{
Data frame with columns:
\describe{
  \item{N}{numeric of group's total observations}
  \item{mean_mpg}{numeric of group's mean mpg}
  \item{mean_hp}{numeric of group's mean horsepower}
  \item{mean_wt}{numeric of group's mean weight}
  \item{mean_disp}{numeric of group's mean displacment}
}
}
\source{
\code{datasets::mtcars}
}
\usage{
car_stats
}
\description{
From \code{datasets::mtcars} produces statistics for N and the means of
"mpg", "hp", "wt", "disp" variables grouped by "cyl".
}
\keyword{datasets}
