% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_waffle_chart.R
\name{create_waffle_chart}
\alias{create_waffle_chart}
\title{create_waffle_chart}
\usage{
create_waffle_chart(
  x = NULL,
  name_col = NULL,
  prop_col = NULL,
  title = NULL,
  background_color = "yellow",
  cell_fill = "purple",
  cell_color = "black",
  cell_nonfill = "white",
  cell_alpha = 1,
  line_width = 1
)
}
\arguments{
\item{x}{Either a dataframe or a vector of named integer proportions.}

\item{name_col}{If *x* is a dataframe, the column name having the variable names.}

\item{prop_col}{If *x* is a dataframe, the column name having the variable proportions.}

\item{title}{A main title for the chart.}

\item{background_color}{A string that sets the background color.}

\item{cell_fill}{A string that sets the cell fill color.}

\item{cell_color}{A string that sets the cell boundary line color.}

\item{cell_nonfill}{A string that sets the non-cell fill color.}

\item{cell_alpha}{A numeric that sets the cell's fill alpha.}

\item{line_width}{A numeric that sets the line width among the cells}
}
\value{
A ggplot class object
}
\description{
Function creates a "waffle" chart that displays percentages for multiple categorical variables.
  For more information on "waffle" charts see \href{https://baryon.be/uncommon-chart-types-waffle-charts/}{What is a waffle chart}.
}
\examples{
library(grid)
library(RplotterPkg)

proportions_v <- c(
  var1=10, var2=40, var3=20, var4=50, var5=5,
  var6=30, var7=10, var8=67, var9=42, var10=33,
  var11=7, var12=35, var13=22, var14=90, var15=43
)

RplotterPkg::create_waffle_chart(
  x = proportions_v,
  title = "Test For 15 Proportion Variables",
  cell_color = "white",
  cell_fill = "brown",
  cell_nonfill = "lightgray",
  cell_alpha = 0.5,
  line_width = 4
)

}
