% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{homeruns_2000}
\alias{homeruns_2000}
\title{homeruns_2000 data}
\format{
Data frame with columns:
\describe{
  \item{YEARS}{numeric years for baseball teams from 1900 to 2000}
  \item{HOMERUNS}{numeric of total homeruns for a baseball team}
}
}
\source{
\href{https://github.com/bayesball/LearnEDAfunctions}{LearnEDAfunctions}
}
\usage{
homeruns_2000
}
\description{
Team home run numbers for different seasons. The
raw tab delimited data (\code{homeruns_2000.txt}) is read by
\code{data.table::fread()} to produce a data frame with numeric
columns for "YEARS" and "HOMERUNS".
}
\keyword{datasets}
