% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kentucky_counties}
\alias{kentucky_counties}
\title{kentucky_counties data}
\format{
An object of class \code{sf} (inherits from \code{data.table}, \code{data.frame}) with 9 rows and 2 columns.
}
\source{
\code{maps::map()}
}
\usage{
kentucky_counties
}
\description{
A simple feature object with the boundary geometries for nine
southeast Kentucky counties.
.

\tabular{ll}{
  \strong{ID} \tab character giving the name of the county \cr
  \strong{geom} \tab MULTIPOLYGON of the Kentucky county boundary
}
}
\keyword{datasets}
