% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{midwest}
\alias{midwest}
\title{midwest data}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 437 rows and 7 columns.
}
\source{
\code{ggplot2::midwest}
}
\usage{
midwest
}
\description{
From the \code{ggplot2::midwest} dataset, selects for midwest counties
percent high school diploma, percent college educated, percentage professional
degree, percent white, percent black, percent asian.

\tabular{ll}{
  \strong{HS_Diploma} \tab numeric percent high school diploma \cr
  \strong{College_Edu} \tab numeric percent college educated \cr
  \strong{Prof_Deg} \tab percentage professional degree \cr
  \strong{White} \tab percent white \cr
  \strong{Black} \tab percent black \cr
  \strong{Asian} \tab percent asian
}
}
\keyword{datasets}
