% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_panel_grid.R
\name{multi_panel_grid}
\alias{multi_panel_grid}
\title{Arranges a group of ggplot2 plots or grid::grobs into multiple panels.}
\usage{
multi_panel_grid(
  layout = NULL,
  col_widths = c(4, 4),
  row_heights = c(4, 4),
  title = NULL,
  subtitle = NULL,
  do_legend = FALSE,
  display_plot = TRUE
)
}
\arguments{
\item{layout}{A named list containing a list for plot objects or grobs \dQuote{plots}, row locations \dQuote{rows} vector,
column locations \dQuote{cols} vector.}

\item{col_widths}{A numeric vector of column widths in inches.}

\item{row_heights}{A numeric vector of row heights in inches.}

\item{title}{A string that sets the title of the figure.}

\item{subtitle}{A string that sets the subtitle of the figure.}

\item{do_legend}{A logical that displays a common legend among the plots.}

\item{display_plot}{A logical that if TRUE will display the plot}
}
\value{
A grob object. Note that if \code{display_plot} is set to \code{FALSE}, then to produce graphical output you must use
\code{grid::grid.draw()} on the returned grob object.
}
\description{
Function creates a figure with title, subtitle, legend and arranges multiple plots
across a given number of rows and columns. The function depends on the \code{grid} and \code{gtable} packages.
}
\author{
Rick Dean
}
