% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_panel_grid.R
\name{multi_panel_grid}
\alias{multi_panel_grid}
\title{Arranges a group of ggplot2 plots or grid::grobs into multiple panels.}
\usage{
multi_panel_grid(
  layout,
  col_widths = c(8, 8),
  row_heights = c(8, 8),
  title = NULL,
  title_fontsz = 14,
  display_plot = TRUE
)
}
\arguments{
\item{layout}{A named list containing a list for plot objects or grobs "plots", row locations "rows" vector,
column locations "cols" vector. This argument is required.}

\item{col_widths}{A numeric vector of each column's widths in centimeters. Note that if there are two
columns 8 centimeters wide then the argument should be \code{c(8,8)}.}

\item{row_heights}{A numeric vector of each row's heights in centimeters. Note that if there are two
rows 10 centimeters high then the argument should be \code{c(10,10)}. If we have one row 10 centimeters
high then this argument should just be assigned the value 10.}

\item{title}{A string that sets the title of the figure.}

\item{title_fontsz}{A numeric that sets the title's font size. The default is 14.}

\item{display_plot}{A logical that if TRUE will display the plot.}
}
\value{
A ggplot2 object if \code{display_plot} is FALSE.
}
\description{
Function creates a figure with title and arranges multiple plots/grobs
across a given number of rows and columns. The function depends on the \code{grid}
and \code{gtable} packages. As an example of defining the \code{layout} argument,
\code{layout} has three named elements: "plots", "rows", and "cols".
"plots" is a list() of ggplot2 plot objects; "rows" defines the row number
for each plot and "cols" defines the column number for each plot. If we have
two plots with both plots on a single row then "rows" = c(1,1) and "cols"
= c(1,2).
}
\author{
Rick Dean
}
