% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{organdata}
\alias{organdata}
\title{organdata data}
\format{
Data frame with columns:
\describe{
  \item{country}{character country name}
  \item{donors}{Organ Donation rate per million population}
}
}
\source{
\code{socviz::organdata}
}
\usage{
organdata
}
\description{
From the \code{socviz::organdata} dataset two variables are
selected: Organ Donation rate per million population and
country.
}
\keyword{datasets}
