% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{penguins_stats}
\alias{penguins_stats}
\title{penguins_stats data}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 3 rows and 5 columns.
}
\source{
\code{palmerpenguins:penguins}
}
\usage{
penguins_stats
}
\description{
The \code{palmerpenguins:penguins} dataset is grouped by the variable
"species" and a data frame is produced with statistics for
average, minimum and maximum body mass.

\tabular{ll}{
  \strong{avg_body_mass} \tab numeric species average body mass \cr
  \strong{min_body_mass} \tab numeric species minimum body mass \cr
  \strong{max_body_mass} \tab numeric species maximum body mass \cr
  \strong{mid_labels} \tab character species average body mass
}
}
\keyword{datasets}
