% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{people}
\alias{people}
\title{people data}
\format{
Data frame with columns:
\describe{
  \item{col_0}{numeric random numbers for ids'}
  \item{col_1}{character first names}
  \item{col_2}{character last names}
  \item{col_3}{numeric ages}
  \item{col_4}{character addresses}
  \item{col_5}{character spouse's first name}
}
}
\source{
R coding with \code{data.table::data.table()}
}
\usage{
people
}
\description{
A made-up data frame with generic column names and mix of
numeric and character data.
}
\keyword{datasets}
