% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{religion}
\alias{religion}
\title{religion data}
\format{
Data frame with columns:
\describe{
  \item{happy}{general happiness}
  \item{religion}{religion in six categories}
}
}
\source{
\code{socviz::gss_sm}
}
\usage{
religion
}
\description{
From the \code{socviz::gss_sm} dataset General Social Survey, 2016,
a sub-data frame was created from the "happy" and "religion" variables.
}
\keyword{datasets}
