% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{states}
\alias{states}
\title{states data}
\format{
Data frame for "South" region:
\describe{
  \item{State}{charactter state name}
  \item{Income}{numeric state income}
  \item{Murder}{numeric state murder rate}
}
}
\source{
\code{datasets::state_name, state_region, state_x77}
}
\usage{
states
}
\description{
Consolidation of data into one dataframe for the "South" region.
}
\keyword{datasets}
